Definitive fix to init drop and config lost when Mango server is not ready. Also a fix to ambiguous update rate set.
